# mes-root/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mes Auto
  description: auto api
  contact:
    email: ixtf1984@gmail.com
  version: 1.0.0
security:
- bearerAuth: []
...
paths:
...
  /api/packageBoxes/{id}/silkCarRecords:
    get:
      operationId: silkCarRecords
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SilkCarRecord'
...
  /api/silks/{id}/silkCarRecords:
    get:
      operationId: silkCarRecords_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SilkCarRecord'
  /api/silks/{id}/packageBox:
    get:
      operationId: packageBox
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBox'
...
  /riamb/SilkDetachEvents:
    post:
      operationId: fetchSilkCarRecord
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Command'
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /riamb/silkCarRecords/codes/{code}:
    get:
      operationId: fetchSilkCarRecord_1
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiambFetchSilkCarRecordResultDTO'
  /riamb/PackageBoxEvents:
    post:
      operationId: fetchSilkCarRecord_2
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Command'
      responses:
        default:
          description: default response
          content:
            application/json: {}
...
components:
  schemas:
...
    Batch:
      required:
      - batchNo
      - id
      - product
      - spec
      - tubeColor
      - workshop
      type: object
      properties:
        id:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        product:
          $ref: '#/components/schemas/Product'
        batchNo:
          type: string
        silkWeight:
          minimum: 1
          type: number
          format: double
        centralValue:
          minimum: 1
          type: number
          format: double
        holeNum:
          minimum: 1
          type: integer
          format: int32
        spec:
          type: string
        tubeColor:
          type: string
        note:
          type: string
    Corporation:
      required:
      - code
      - id
      - name
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
    DyeingResult:
      required:
      - createDateTime
      - creator
      - dateTime
      - id
      - lineMachine
      - silk
      type: object
      properties:
        id:
          type: string
        silk:
          $ref: '#/components/schemas/Silk'
        lineMachine:
          $ref: '#/components/schemas/LineMachine'
        spindle:
          minimum: 1
          type: integer
          format: int32
        dateTime:
          type: string
          format: date-time
        hasException:
          type: boolean
        grade:
          $ref: '#/components/schemas/Grade'
        silkExceptions:
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        silkNotes:
          type: array
          items:
            $ref: '#/components/schemas/SilkNote'
        creator:
          $ref: '#/components/schemas/Operator'
        createDateTime:
          type: string
          format: date-time
        modifier:
          $ref: '#/components/schemas/Operator'
        modifyDateTime:
          type: string
          format: date-time
        submitted:
          type: boolean
        formConfigValueData:
          $ref: '#/components/schemas/JsonNode'
        formConfig:
          $ref: '#/components/schemas/JsonNode'
    DyeingResultInfo:
      type: object
      properties:
        dyeingResult:
          $ref: '#/components/schemas/DyeingResult'
        grade:
          $ref: '#/components/schemas/Grade'
...
    Grade:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        sortBy:
          type: integer
          format: int32
    JsonNode:
      type: object
      properties:
        valueNode:
          type: boolean
        containerNode:
          type: boolean
        missingNode:
          type: boolean
        object:
          type: boolean
        pojo:
          type: boolean
        integralNumber:
          type: boolean
        floatingPointNumber:
          type: boolean
        textual:
          type: boolean
        boolean:
          type: boolean
        binary:
          type: boolean
        int:
          type: boolean
        number:
          type: boolean
        bigDecimal:
          type: boolean
        bigInteger:
          type: boolean
        double:
          type: boolean
        long:
          type: boolean
        float:
          type: boolean
        nodeType:
          type: string
          enum:
          - ARRAY
          - BINARY
          - BOOLEAN
          - MISSING
          - NULL
          - NUMBER
          - OBJECT
          - POJO
          - STRING
        short:
          type: boolean
        array:
          type: boolean
        null:
          type: boolean
    Line:
      required:
      - id
      - name
      - workshop
      type: object
      properties:
        id:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        name:
          type: string
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
    LineMachine:
      required:
      - id
      - line
      - spindleSeq
      type: object
      properties:
        id:
          type: string
        line:
          $ref: '#/components/schemas/Line'
        item:
          minimum: 1
          type: integer
          format: int32
        spindleNum:
          minimum: 1
          type: integer
          format: int32
        spindleSeq:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            type: integer
            format: int32
    Operator:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        hrId:
          type: string
        oaId:
          type: string
        phone:
          type: string
        admin:
          type: boolean
    Product:
      required:
      - code
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
    SapT001l:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        lgort:
          type: string
        lgobe:
          type: string
    Silk:
      required:
      - batch
      - code
      - doffingDateTime
      - doffingOperator
      - doffingType
      - grade
      - id
      - lineMachine
      - spindle
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        doffingNum:
          type: string
        doffingOperator:
          $ref: '#/components/schemas/Operator'
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
        doffingDateTime:
          type: string
          format: date-time
        lineMachine:
          $ref: '#/components/schemas/LineMachine'
        spindle:
          type: integer
          format: int32
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        exception:
          $ref: '#/components/schemas/SilkException'
        exceptions:
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        dyeingExceptionStrings:
          type: array
          items:
            type: string
        dyeingSample:
          type: boolean
        detached:
          type: boolean
        packageDateTime:
          type: string
          format: date-time
    SilkCar:
      required:
      - code
      - id
      - number
      - type
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - DEFAULT
          - BIG_SILK_CAR
        number:
          type: string
        code:
          type: string
        row:
          minimum: 1
          type: integer
          format: int32
        col:
          minimum: 1
          type: integer
          format: int32
    SilkCarRecord:
      required:
      - batch
      - grade
      - id
      type: object
      properties:
        id:
          type: string
        silkCar:
          $ref: '#/components/schemas/SilkCar'
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        doffingOperator:
          $ref: '#/components/schemas/Operator'
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
        doffingDateTime:
          type: string
          format: date-time
        carpoolOperator:
          $ref: '#/components/schemas/Operator'
        carpoolDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        startDateTime:
          type: string
          format: date-time
        initCommand:
          $ref: '#/components/schemas/JsonNode'
        initSilks:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntime'
...
    SilkException:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    SilkNote:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mustFeedback:
          type: boolean
    SilkRuntime:
      type: object
      properties:
        sideType:
          type: string
          enum:
          - A
          - B
        row:
          type: integer
          format: int32
        col:
          type: integer
          format: int32
        silk:
          $ref: '#/components/schemas/Silk'
        exceptions:
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/SilkNote'
        grade:
          $ref: '#/components/schemas/Grade'
        firstDyeingResultInfo:
          $ref: '#/components/schemas/DyeingResultInfo'
        crossDyeingResultInfo:
          $ref: '#/components/schemas/DyeingResultInfo'
        multiDyeingResultInfo:
          $ref: '#/components/schemas/DyeingResultInfo'
    Workshop:
      required:
      - code
      - corporation
      - id
      - name
      type: object
      properties:
        id:
          type: string
        corporation:
          $ref: '#/components/schemas/Corporation'
        name:
          type: string
        code:
          maxLength: 1
          minLength: 1
          type: string
        note:
          type: string
        sapT001ls:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
        sapT001lsForeign:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
        sapT001lsPallet:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
...
    PackageBox:
      required:
      - batch
      - code
      - grade
      - id
      - type
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - AUTO
          - MANUAL
          - MANUAL_APPEND
          - WORK_FETCH
          - FOREIGN
        code:
          type: string
        saleType:
          type: string
          enum:
          - DOMESTIC
          - FOREIGN
        palletCode:
          type: string
        palletType:
          type: string
        packageType:
          type: string
        foamType:
          type: string
        foamNum:
          type: integer
          format: int32
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        silkCount:
          minimum: 1
          type: integer
          format: int32
        grossWeight:
          minimum: 1
          type: number
          format: double
        netWeight:
          minimum: 1
          type: number
          format: double
        riambJobId:
          type: string
        printDate:
          type: string
          format: date-time
        printClass:
          $ref: '#/components/schemas/PackageClass'
        printCount:
          minimum: 0
          type: integer
          format: int32
        automaticPackeLine:
          type: string
        budat:
          type: string
          format: date-time
        budatClass:
          $ref: '#/components/schemas/PackageClass'
        sapT001l:
          $ref: '#/components/schemas/SapT001l'
        inWarehouse:
          type: boolean
        creator:
          $ref: '#/components/schemas/Operator'
    PackageClass:
      required:
      - id
      - name
      - riambCode
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        riambCode:
          type: string
        sortBy:
          type: integer
          format: int32
...
    SilkInfo:
      required:
      - code
      type: object
      properties:
        code:
          type: string
        weight:
          type: number
...
    Command:
      required:
      - code
      - createDateTime
      - grossWeight
      - jobInfo
      - netWeight
      - silkInfos
      type: object
      properties:
        code:
          type: string
        jobInfo:
          $ref: '#/components/schemas/AutomaticPackeJobInfo'
        netWeight:
          minimum: 1
          type: number
        grossWeight:
          minimum: 1
          type: number
        silkCount:
          minimum: 1
          type: integer
          format: int32
        createDateTime:
          type: string
          format: date-time
        silkInfos:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SilkInfo'
        palletCode:
          type: string
        otherInfo:
          type: string
    SilkCarInfo:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        row:
          type: integer
          format: int32
        col:
          type: integer
          format: int32
        batchNo:
          type: string
    RiambFetchSilkCarRecordResultDTO:
      type: object
      properties:
        packeFlage:
          type: string
        silkCarInfo:
          $ref: '#/components/schemas/SilkCarInfo'
        silkCount:
          type: integer
          format: int32
        silkInfos:
          type: array
          items:
            $ref: '#/components/schemas/SilkInfo'
    AutomaticPackeJobInfo:
      required:
      - automaticPackeLine
      - batchNo
      - createDateTime
      - creatorHrId
      - gradeName
      - id
      type: object
      properties:
        id:
          type: string
        automaticPackeLine:
          type: string
        batchNo:
          type: string
        gradeName:
          type: string
        packageClassNo:
          type: string
        saleType:
          type: string
        budatDate:
          type: string
          format: date-time
        palletType:
          type: string
        packageType:
          type: string
        foamType:
          type: string
        foamNum:
          type: integer
          format: int32
        creatorHrId:
          type: string
        createDateTime:
          type: string
          format: date-time
        lgort:
          type: string
        otherInfo:
          type: string
...